SystemUI: Initialize QS tiles in inactive state
Now that the QS fragment is recreated when changing the UI mode (so that
it follows light/dark themes), all tiles flash with active color briefly
because the new views become visible before states are refreshed.

Initializing tiles in the inactive state is much less disruptive, and
the effect is very hard to see as compared to the active color because
the background color is much less prominent.

Change-Id: I048171d503f5533e91bab486b8805ac15c329f31
Signed-off-by: HeroBuxx <herobuxx@conqueros.co>